+++ core/network/src/main/java/team/retum/network/util/TokenInterceptor.kt
@@ -1,4 +1,4 @@
-package team.retum.network
+package team.retum.network.util
 
 import okhttp3.Interceptor
 import okhttp3.Response
